resolved conflicts

+++ src/main/resources/db/migration/V3__ADELIN_link_players_to_clubs.sql
@@ -1,7 +1,7 @@
 Use football_club_simulator;
 
 alter table players
-    add current_club_id int null;
+    add current_club_id BIGINT null;
 
 alter table players
     add constraint current_club_id
